Register startbootstrap-sb-admin in main and names refactor
