+++ netedu-app/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <title>Net Education</title>
   </head>
-  <body>
+  <body id="page-top" class="fixed-nav sticky-footer bg-dark">
     <div id="app"></div>
     <!-- Static jQuery scripts -->
     <script src="static/jquery/jquery.min.js"></script>
