+++ netedu-app/src/App.vue
@@ -1,25 +1,25 @@
 <template>
   <div id="app" class="container-fluid">
-    <AppContentComp/>
+    <AppContent/>
   </div>
 </template>
 
 <script>
-import AppContentComp from './components/AppComps/AppContentComp'
+import AppContent from './components/App/AppContent'
 export default {
   name: 'App',
   components: {
-    AppContentComp
+    AppContent
   }
 }
 </script>
 
 <style lang="scss" scoped>
-#app {
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-font-smoothing: antialiased;
-  color: #2c3e50;
-  font-family: 'Avenir', Helvetica, Arial, sans-serif;
-  text-align: center;
-}
+// #app {
+//   -moz-osx-font-smoothing: grayscale;
+//   -webkit-font-smoothing: antialiased;
+//   color: #2c3e50;
+//   font-family: 'Avenir', Helvetica, Arial, sans-serif;
+//   text-align: center;
+// }
 </style>
