+++ netedu-app/src/components/App/AppContent.vue
