+++ netedu-app/src/components/App/app.config
